pyproject.toml
[project]
name = "dsa-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "discord-py>=2.7.1",
    "python-dotenv>=1.0.1",
]
